+++ R/stats19-package.R
@@ -1,7 +1,8 @@
 .onAttach = function(libname, pkgname) {
   msg = paste0(
-      "Data provided under OGL v3.0. Cite the source and link to:\n",
-      "www.nationalarchives.gov.uk/doc/open-government-licence/version/3/"
-    )
+    "Data provided under OGL v3.0.\n",
+    "To cite stats19 in publications: Lovelace et al. (2019) JOSS <doi:10.21105/joss.01181>.\n",
+    "Run citation(\"stats19\") for BibTeX entry."
+  )
   packageStartupMessage(msg)
 }